Use FileUtils instead of function in Scanner
