+++ tests/lib/Files/Cache/ScannerTest.php
@@ -10,6 +10,7 @@ namespace Test\Files\Cache;
 
 use OC\Files\Cache\CacheEntry;
 use OC\Files\Storage\Storage;
+use OC\Files\Utils\FileUtils;
 use OCA\Files_Sharing\ISharedStorage;
 use OCP\Files\IHomeStorage;
 use OCP\Files\Storage\ILockingStorage;
@@ -323,28 +324,6 @@ class ScannerTest extends \Test\TestCase {
 
 		$cachedData = $this->cache->get('folder/bar.txt');
 		$this->assertEquals($newFolderId, $cachedData['parent']);
-	}
-
-	/**
-	 * @dataProvider dataTestIsPartialFile
-	 *
-	 * @param string $path
-	 * @param bool $expected
-	 */
-	public function testIsPartialFile($path, $expected) {
-		$this->assertSame($expected,
-			$this->scanner->isPartialFile($path)
-		);
-	}
-
-	public function dataTestIsPartialFile() {
-		return [
-			['foo.txt.part', true],
-			['/sub/folder/foo.txt.part', true],
-			['/sub/folder.part/foo.txt', true],
-			['foo.txt', false],
-			['/sub/folder/foo.txt', false],
-		];
 	}
 
 	public function failGetDataProvider() {
